Refactored DTOs and Interfaces in Chirp.Core
