+++ src/Chirp.Core/DTOs/AuthorDTO.cs
@@ -1,0 +1,7 @@
+﻿namespace Chirp.Core.DTOs;
+
+public class AuthorDTO
+{
+    public required string Name { get; set; }
+    public required string Email { get; set; }
+}
