+++ src/Chirp.Core/DTOs/CheepDTO.cs
@@ -1,0 +1,8 @@
+﻿namespace Chirp.Core.DTOs;
+
+public class CheepDTO
+{
+    public required string AuthorName { get; set; } // Author's name
+    public required string Text { get; set; } // Message text
+    public required string FormattedTimeStamp { get; set; } // Time stamp as a formatted string
+}
