+++ src/Chirp.Core/Interfaces/ICheepRepository.cs
@@ -1,0 +1,12 @@
+﻿using Chirp.Core.DTOs;
+
+namespace Chirp.Core.Interfaces;
+
+public interface ICheepRepository
+{
+    Task<List<CheepDTO>> ReadCheepsFromAuthor(string userName, int page);
+    Task<List<CheepDTO>> ReadAllCheeps(int page);
+    Task CreateCheep(CheepDTO newCheep);
+    Task UpdateCheep(CheepDTO alteredCheep);
+    Task<int> GetTotalPages(string authorName);
+}
