+++ src/Chirp.Infrastructure/Services/CheepService.cs
@@ -3,8 +3,8 @@ using Chirp.Infrastructure.Repositories;
 namespace Chirp.Infrastructure.Services;
 public interface ICheepService
 {
-    public Task<List<CheepDTO>> GetCheeps(int page);
-    public Task<List<CheepDTO>> GetCheepsFromAuthor(string author, int page);
+    public Task<List<Core.DTOs.CheepDTO>> GetCheeps(int page);
+    public Task<List<Core.DTOs.CheepDTO>> GetCheepsFromAuthor(string author, int page);
     
     public Task<int> GetTotalPageNumber(string authorName);
 }
@@ -20,28 +20,13 @@ public class CheepService : ICheepService
     {
         return await _cheepRepository.GetTotalPages(authorName);
     }
-    public async Task<List<CheepDTO>> GetCheeps(int page)
+    public async Task<List<Core.DTOs.CheepDTO>> GetCheeps(int page)
     {
         return await _cheepRepository.ReadAllCheeps(page);
     }
     
-    public async Task<List<CheepDTO>> GetCheepsFromAuthor(string author, int page)
+    public async Task<List<Core.DTOs.CheepDTO>> GetCheepsFromAuthor(string author, int page)
     {
         return await _cheepRepository.ReadCheepsFromAuthor(author, page);
     }
-}
-
-// Data Transfer Object for Cheep
-public class CheepDTO
-{
-    public required string AuthorName { get; set; } // Author's name
-    public required string Text { get; set; } // Message text
-    public required string FormattedTimeStamp { get; set; } // Time stamp as a formatted string
-}
-
-// Data Transfer Object for Author
-public class AuthorDTO
-{
-    public required string Name { get; set; }
-    public required string Email { get; set; }
 }
